(PC-31331)[PRO] refactor: split individual and collective offers url computation
