+++ pro/src/core/Offers/utils/__specs__/computeCollectiveOffersUrl.spec.ts
@@ -1,0 +1,28 @@
+import { CollectiveOfferDisplayedStatus } from 'apiClient/v1'
+
+import { computeCollectiveOffersUrl } from '../computeCollectiveOffersUrl'
+
+describe('computeCollectiveOffersUrl', () => {
+  it('should build proper query given collective offers filters', () => {
+    const offersSearchFilters = {
+      name: 'test',
+      offererId: 'AY',
+      venueId: 'EQ',
+      categoryId: 'CINEMA',
+      status: [
+        CollectiveOfferDisplayedStatus.ACTIVE,
+        CollectiveOfferDisplayedStatus.EXPIRED,
+      ],
+      creationMode: 'manual',
+      periodBeginningDate: '2020-11-30T00:00:00+01:00',
+      periodEndingDate: '2021-01-07T23:59:59+01:00',
+      page: 2,
+    }
+
+    const value = computeCollectiveOffersUrl(offersSearchFilters)
+
+    expect(value).toBe(
+      '/offres/collectives?page=2&nom=test&structure=AY&lieu=EQ&categorie=CINEMA&statut=active&statut=expiree&creation=manuelle&periode-evenement-debut=2020-11-30T00%3A00%3A00%2B01%3A00&periode-evenement-fin=2021-01-07T23%3A59%3A59%2B01%3A00'
+    )
+  })
+})
